renamed context.json for all the extensions in pojos and added valuators in pojos
