+++ certgen/html-generator/src/main/java/org/incredible/HTMLGenerator.java
@@ -9,9 +9,9 @@ import org.slf4j.LoggerFactory;
 import org.apache.commons.io.FileUtils;
 
 
-public class HTMLTemplateGenerator {
+public class HTMLGenerator {
 
-    private static Logger logger = LoggerFactory.getLogger(HTMLTemplateGenerator.class);
+    private static Logger logger = LoggerFactory.getLogger(HTMLGenerator.class);
     private static final String HTML_TEMPLATE_NAME = "template.html";
 
     public void generateTemplate(Assertion assertion) {
@@ -34,6 +34,7 @@ public class HTMLTemplateGenerator {
         htmlString = htmlString.replace("$recipient", assertion.getRecipient().getName());
         htmlString = htmlString.replace("$img", path);
         htmlString = htmlString.replace("$course", assertion.getBadge().getName());
+        htmlString = htmlString.replace("$dated", assertion.getIssuedOn());
         File newHtmlFile = new File(id + ".html");
         try {
             FileUtils.writeStringToFile(newHtmlFile, htmlString);
@@ -44,7 +45,7 @@ public class HTMLTemplateGenerator {
 
 
     private static String getPath(String file) {
-        ClassLoader loader = HTMLTemplateGenerator.class.getClassLoader();
+        ClassLoader loader = HTMLGenerator.class.getClassLoader();
 
         String result = null;
         try {
