+++ certgen/pojos/src/main/java/org/incredible/pojos/ob/Assertion.java
@@ -1,7 +1,13 @@
 package org.incredible.pojos.ob;
 
 
+import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import org.incredible.pojos.CompositeIdentityObject;
+import org.incredible.pojos.ob.valuator.ExpiryDateValuator;
+import org.incredible.pojos.ob.valuator.IssuedDateValuator;
+
+import java.text.MessageFormat;
+import java.text.ParseException;
 
 
 /**
@@ -84,7 +90,16 @@ public class Assertion extends OBBase {
     }
 
     public void setIssuedOn(String issuedOn) {
-        this.issuedOn = issuedOn;
+
+        IssuedDateValuator issuedDateValuator = new IssuedDateValuator();
+        try {
+            if (issuedDateValuator.evaluates(issuedOn) == null) {
+                throw new IllegalArgumentException("Issued date is not in given format");
+            } else
+                this.issuedOn = issuedDateValuator.evaluates(issuedOn);
+        } catch (IllegalArgumentException e) {
+        }
+
     }
 
     public CompositeIdentityObject getRecipient() {
@@ -124,7 +139,19 @@ public class Assertion extends OBBase {
     }
 
     public void setExpires(String expires) {
-        this.expires = expires;
+        try {
+            if(this.getIssuedOn() == null) {
+                throw new Exception("Issued date is not in format");
+            } else {
+                ExpiryDateValuator valuator = new ExpiryDateValuator(this.getIssuedOn());
+                if (valuator.evaluates(expires) == null) {
+                    throw new IllegalArgumentException("Expiry date is in wrong format");
+                } else
+                    this.expires = valuator.evaluates(expires);
+            }
+        } catch (Exception e) {
+        }
+
     }
 
     public VerificationObject getVerification() {
