+++ certgen/pojos/src/main/java/org/incredible/pojos/ob/valuator/ExpiryDateValuator.java
@@ -1,4 +1,4 @@
-package org.incredible.certProcessor;
+package org.incredible.pojos.ob.valuator;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
