+++ certgen/pojos/src/main/resources/extensions/AssessedEvidence/schema.json
