+++ certgen/pojos/src/main/resources/extensions/Assessment/schema.json
