+++ certgen/pojos/src/main/resources/extensions/CertificateExtension/schema.json
