+++ certgen/pojos/src/main/resources/extensions/CompositeIdentityObject/context.json
