+++ certgen/pojos/src/main/resources/extensions/MarksAssessment/schema.json
