+++ certgen/pojos/src/main/resources/extensions/RankAssessment/schema.json
