+++ certgen/pojos/src/main/resources/extensions/SignatoryExtension/schema.json
